[Cache] optimize Cache implementation
Use code generation & memoization to generate functions specific to any
number of path entries. Also add a get/setter that can laziliy
initialize cache values.

Disclaimer: I modified only it's use in the Unit module, dunno if there
are more places.

On a completely unscientific benchmark doing

  local t = os.clock();
  for i = 1, 100000 do
    local path = "X."..i..".Y";
    local b = Cache.Get(path) or Cache.Set(path, i);
  end
  print(os.clock() - t);

for the current implementation and

  local t = os.clock();
  for i = 1, 100000 do
    local b = Cache.Get("X", i, "Y", function() return i end);
  end
  print(os.clock() - t);

for the new one reduces the final printed time from .832 to .045 which
is ~18 times faster.