[Unit & Player] 'cache' unitInfo into locals
The "initialization" of the variable (if we don't have a table for this
guid yet) copy-pasted everywhere is obviously ugly, but requires to be
inlined if we care about speed here.

There are basically 2 ways to remove the ugliness while retaining the
speed:
 * do the check & creation inside :GUID() as nothing ever tries to get
   at the unit info table before calling it
 * setmetatable() on Cache.UnitInfo with an __index that simply creates
   tables if there is nothing at the given key
The former can be even sped up slightly more by caching the table on the
unit table itself (and niling it for all units inside the "tick").